Semi Final attempt at debugging WaveSystem before total replacement
Enemies are spawning but no do shoot and do not trigger BattleOver

+++ Assets/Scripts/UI Scripts/ResultsKeeper.cs
@@ -32,7 +32,7 @@ public class ResultsKeeper : MonoBehaviour
 
     public int GetWaveInfo()
     { 
-        return (int)GlobalSettings.globalWaveCount;
+        return GlobalSettings.globalWaveCount;
     }
 
     public void ShowResults()
